barcode quantity类型改为int

+++ warehouse-operation-api/src/main/java/com/baozun/scm/primservice/whoperation/model/sku/SkuBarcode.java
@@ -46,7 +46,7 @@ public class SkuBarcode extends BaseModel {
     /** 操作人ID */
     private Long modifiedId;
     /** 数量 */
-    private Long quantity;
+    private Integer quantity;
     /**
      * 1.可用;2.已删除;0.禁用
      */
@@ -110,11 +110,11 @@ public class SkuBarcode extends BaseModel {
         this.modifiedId = modifiedId;
     }
 
-    public Long getQuantity() {
+    public Integer getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Long quantity) {
+    public void setQuantity(Integer quantity) {
         this.quantity = quantity;
     }
